reorganized serializer-view-url code.
